LayoutTests/fast/page-color-sampling/color-sampling-ignores-non-fixed-content.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
        body {
            font-family: system-ui;
            margin: 0;
        }

        header {
            text-align: center;
            position: fixed;
            top: 0px;
            left: 0;
            font-size: 32px;
            width: 100%;
            height: 250px;
            background: rgba(255, 100, 0, 0.8);
        }

        .top {
            width: 100%;
            height: 40px;
            background: black;
            position: absolute;
            top: 0;
        }

        .tall {
            width: 100%;
            height: 2000px;
        }
    </style>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        await UIHelper.setObscuredInsets(100, 0, 0, 0);
        await UIHelper.ensurePresentationUpdate();
        colorBeforeFading = await UIHelper.fixedContainerEdgeColors();

        document.querySelector("header").style.opacity = 0.01;
        await UIHelper.ensurePresentationUpdate();
        colorAfterFading = await UIHelper.fixedContainerEdgeColors();

        shouldBeEqualToString("colorBeforeFading.top", "rgba(255, 100, 0, 0.8)");
        shouldBeEqualToString("colorAfterFading.top", "rgba(255, 100, 0, 0.8)");
        finishJSTest();
    });
    </script>
</head>
<body>
<div class="top"></div>
<div class="tall"></div>
<header></header>
</body>
</html>
