LayoutTests/fast/page-color-sampling/color-sampling-ignores-non-fixed-content-expected.txt

PASS colorBeforeFading.top is "rgba(255, 100, 0, 0.8)"
PASS colorAfterFading.top is "rgba(255, 100, 0, 0.8)"
PASS successfullyParsed is true

TEST COMPLETE
